Replace old modules, rebundle.
